Tested CATS mode on RS41. Add some CATS documentation in README. Add advanced GPS configuration options to config.h for more frequent position updates, which are useful especially when using CATS. Format code. Make sure radio main loop handlers do not execute code when transmission is already finished. Add message/comment text variables for debugging transmissions: data counter and GPS update status.

+++ src/codecs/cats/cats.c
@@ -17,9 +17,8 @@ cats_packet cats_create(uint8_t *data)
 // the packet after calling!
 size_t cats_fully_encode(cats_packet packet, uint8_t *out)
 {
-
     // 0. CRC
-    int new_len = cats_append_crc(packet.data, packet.len);
+    size_t new_len = cats_append_crc(packet.data, packet.len);
     // 1. whiten
     cats_whiten(packet.data, new_len);
     // 2. ldpc
